rename files and reorg utilities
