+++ featuretools/primitives/standard/aggregation_primitives/all_primitive.py
