+++ featuretools/primitives/standard/aggregation_primitives/min_primitive.py
