+++ featuretools/primitives/standard/transform_primitives/datetime_transform_primitives/is_federal_holiday.py
@@ -3,7 +3,9 @@ from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import BooleanNullable, Datetime
 
 from featuretools.primitives.base import TransformPrimitive
-from featuretools.primitives.utils import HolidayUtil
+from featuretools.primitives.standard.transform_primitives.datetime_transform_primitives.utils import (
+    HolidayUtil,
+)
 
 
 class IsFederalHoliday(TransformPrimitive):
