+++ featuretools/tests/primitive_tests/test_rolling_primitive.py
@@ -10,7 +10,7 @@ from featuretools.primitives import (
     RollingSTD,
     RollingTrend,
 )
-from featuretools.primitives.standard.transform_primitives.time_series_primitives.rolling_primitive_utils import (
+from featuretools.primitives.standard.transform_primitives.time_series_primitives.utils import (
     apply_rolling_agg_to_series,
 )
 from featuretools.tests.primitive_tests.utils import get_number_from_offset
